+++ relay-metrics/src/view.rs
@@ -2,10 +2,13 @@ use relay_common::time::UnixTimestamp;
 use serde::ser::{SerializeMap, SerializeSeq};
 use serde::Serialize;
 
-use crate::{aggregator, CounterType, DistributionType, GaugeValue, SetType, SetValue};
+use crate::{
+    aggregator, BucketMetadata, CounterType, DistributionType, GaugeValue, SetType, SetValue,
+};
 use std::collections::BTreeMap;
 use std::fmt;
 use std::ops::Range;
+use std::sync::Arc;
 
 use crate::bucket::Bucket;
 use crate::BucketValue;
@@ -374,6 +377,13 @@ impl<'a> BucketView<'a> {
         &self.inner.name
     }
 
+    /// Returns the name of the bucket.
+    ///
+    /// Caller holds shared ownership of the string.
+    pub fn clone_name(&self) -> Arc<str> {
+        Arc::clone(&self.inner.name)
+    }
+
     /// Value of the bucket view.
     pub fn value(&self) -> BucketViewValue<'a> {
         match &self.inner.value {
@@ -396,6 +406,13 @@ impl<'a> BucketView<'a> {
     /// See also: [`Bucket::tag()`]
     pub fn tag(&self, name: &str) -> Option<&'a str> {
         self.inner.tag(name)
+    }
+
+    /// Returns the metadata for this bucket.
+    ///
+    /// See also: [`Bucket::metadata`].
+    pub fn metadata(&self) -> &BucketMetadata {
+        &self.inner.metadata
     }
 
     /// Number of raw datapoints in this view.
@@ -514,6 +531,7 @@ impl<'a> Serialize for BucketView<'a> {
             name,
             value: _,
             tags,
+            metadata,
         } = self.inner;
 
         let len = match tags.is_empty() {
@@ -538,6 +556,9 @@ impl<'a> Serialize for BucketView<'a> {
 
         if !tags.is_empty() {
             state.serialize_entry("tags", tags)?;
+        }
+        if !metadata.is_default() {
+            state.serialize_entry("metadata", metadata)?;
         }
 
         state.end()