+++ relay-server/src/envelope.rs
@@ -578,6 +578,8 @@ fn is_true(value: &bool) -> bool {
 pub struct SourceQuantities {
     /// Transaction quantity.
     pub transactions: usize,
+    /// Spans quantity.
+    pub spans: usize,
     /// Profile quantity.
     pub profiles: usize,
     /// Total number of buckets.
@@ -586,8 +588,16 @@ pub struct SourceQuantities {
 
 impl AddAssign for SourceQuantities {
     fn add_assign(&mut self, other: Self) {
-        self.transactions += other.transactions;
-        self.profiles += other.profiles;
+        let Self {
+            transactions,
+            spans,
+            profiles,
+            buckets,
+        } = self;
+        *transactions += other.transactions;
+        *spans += other.spans;
+        *profiles += other.profiles;
+        *buckets += other.buckets;
     }
 }